requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.65.0
python-dotenv>=0.19.0
